html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-move-processed-source-to-end.html

<!doctype html>
<title>pointer updates (moveBefore processed source to the end)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
var events = [];
var s1 = 0;
var s2 = 0;
var moved = false;

function s2_error() {
  s2++;
  events.push('s2');
  if (moved) {
    return;
  }
  moved = true;
  var video = document.getElementById('v');
  var source1 = document.getElementById('source-s1');
  source1.src = 'resources/delayed-broken-video.py?move-processed-source-to-end';
  video.moveBefore(source1, null);
}
</script>
<video id="v"><source id="source-s1" onerror="s1++; events.push('s1')"><source id="source-s2" onerror="s2_error()" src="resources/delayed-broken-video.py"></video>
<script>
async_test(function(t) {
  assert_implements(typeof document.getElementById('v').moveBefore == 'function', 'ParentNode.moveBefore() is required for this test');

  window.onload = t.step_func(function() {
    assert_equals(s1, 2, 'error events on s1');
    assert_equals(s2, 1, 'error events on s2');
    assert_array_equals(events, ['s1', 's2', 's1'], 'processed source order');
    t.done();
  });
});
</script>
